# Conflicts:
#	Vagrantfile
#	src/Client/SDL/enemy.cpp
#	src/Client/Threads/senderThread.cpp
#	src/Server/Game/characters/character.cpp
#	src/Server/Game/characters/character.h
#	src/Server/Game/characters/states/damage.cpp
#	src/Server/Game/characters/states/dead.cpp
#	src/Server/Game/characters/states/idle.cpp
#	src/Server/Game/characters/states/jumping.cpp
#	src/Server/Game/characters/states/jumping.h
#	src/Server/Game/characters/states/moving.cpp
#	src/Server/Game/characters/states/moving.h
#	src/Server/Game/characters/states/shooting.cpp
#	src/Server/Game/entityFactory.cpp
#	src/Server/Game/entityFactory.h
#	src/Server/Game/gameMap.cpp
#	src/Server/Game/gameMap.h
#	src/Server/Game/weapons/blaster.cpp

+++ src/Server/Game/entityFactory.cpp
@@ -1,4 +1,5 @@
 #include "entityFactory.h"
+
 #include "characters/jazz.h"
 #include "characters/lori.h"
 #include "characters/spaz.h"
@@ -6,10 +7,12 @@
 #include "enemies/jumpingEnemy.h"
 #include "enemies/walkingEnemy.h"
 
+#include "gameMap.h"
+
 EntityFactory::EntityFactory(GameMap& gameMap): gameMap(gameMap) {}
 
 std::shared_ptr<Character> EntityFactory::createCharacter(uint8_t playerId, CharacterType type,
-                                                          Vector<uint8_t> position) {
+                                                          Vector<uint32_t> position) {
     switch (type) {
         case CharacterType::JAZZ:
             return createJazz(playerId, position);
@@ -22,7 +25,7 @@ std::shared_ptr<Character> EntityFactory::createCharacter(uint8_t playerId, Char
 }
 
 std::shared_ptr<Enemy> EntityFactory::createEnemy(uint8_t enemyId, EnemyType type,
-                                                  Vector<uint8_t> position) {
+                                                  Vector<uint32_t> position) {
     switch (type) {
         case EnemyType::TURTLE:
             return createWalker(enemyId, position);
@@ -34,46 +37,26 @@ std::shared_ptr<Enemy> EntityFactory::createEnemy(uint8_t enemyId, EnemyType typ
     return nullptr;
 }
 
-// std::shared_ptr<Obstacle> EntityFactory::createObstacle(const std::string& type, Vector<uint8_t> position) {
-//     return std::make_shared<Obstacle>(type, position);
-// }
-
-// std::shared_ptr<GoldCoin> EntityFactory::createGoldCoin(Vector<uint8_t> position) {
-//     return std::make_shared<GoldCoin>(gameMap, position);
-// }
-
-// std::shared_ptr<SilverCoin> EntityFactory::createSilverCoin(Vector<uint8_t> position) {
-//     return std::make_shared<SilverCoin>(gameMap, position);
-// }
-
-// std::shared_ptr<Gem> EntityFactory::createGem(Vector<uint8_t> position) {
-//     return std::make_shared<Gem>(gameMap, position);
-// }
-
-// std::shared_ptr<Tile> EntityFactory::createTile(const std::string& type, Vector<uint8_t> position) {
-//     return std::make_shared<Tile>(type, position);
-// }
-
-std::shared_ptr<Character> EntityFactory::createJazz(uint8_t playerId, Vector<uint8_t> position) {
+std::shared_ptr<Character> EntityFactory::createJazz(uint8_t playerId, Vector<uint32_t> position) {
     return std::make_shared<Jazz>(gameMap, position, playerId);
 }
 
-std::shared_ptr<Character> EntityFactory::createSpaz(uint8_t playerId, Vector<uint8_t> position) {
+std::shared_ptr<Character> EntityFactory::createSpaz(uint8_t playerId, Vector<uint32_t> position) {
     return std::make_shared<Spaz>(gameMap, position, playerId);
 }
 
-std::shared_ptr<Character> EntityFactory::createLori(uint8_t playerId, Vector<uint8_t> position) {
+std::shared_ptr<Character> EntityFactory::createLori(uint8_t playerId, Vector<uint32_t> position) {
     return std::make_shared<Lori>(gameMap, position, playerId);
 }
 
-std::shared_ptr<Enemy> EntityFactory::createWalker(uint8_t enemyId, Vector<uint8_t> position) {
+std::shared_ptr<Enemy> EntityFactory::createWalker(uint8_t enemyId, Vector<uint32_t> position) {
     return std::make_shared<WalkingEnemy>(gameMap, position, enemyId);
 }
 
-std::shared_ptr<Enemy> EntityFactory::createFlyer(uint8_t enemyId, Vector<uint8_t> position) {
+std::shared_ptr<Enemy> EntityFactory::createFlyer(uint8_t enemyId, Vector<uint32_t> position) {
     return std::make_shared<FlyingEnemy>(gameMap, position, enemyId);
 }
 
-std::shared_ptr<Enemy> EntityFactory::createJumper(uint8_t enemyId, Vector<uint8_t> position) {
+std::shared_ptr<Enemy> EntityFactory::createJumper(uint8_t enemyId, Vector<uint32_t> position) {
     return std::make_shared<JumpingEnemy>(gameMap, position, enemyId);
 }
